requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
jinja2
python-multipart